backend/requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.111.0
h11==0.14.0
idna==3.4
pydantic==2.7.3
python-dotenv==0.21.0
PyYAML==6.0.2
sniffio==1.3.0
starlette==0.37.2
typing_extensions==4.12.2
uvicorn==0.30.1
watchfiles==0.18.1
websockets==10.4
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
email-validator==2.0.0
sqlalchemy
psycopg2-binary
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0